cleanup dead code, fix docs, fixate cmlxc version
