+++ cmdeploy/src/cmdeploy/tests/plugin.py
@@ -1,5 +1,4 @@
 import imaplib
-import ipaddress
 import itertools
 import os
 import random
@@ -10,18 +9,17 @@ import time
 from pathlib import Path
 
 import pytest
-from chatmaild.config import read_config, format_mail_domain, is_valid_ipv4
+from chatmaild.config import is_valid_ipv4, read_config
+from domain_validator import DomainValidator
 
+
+def format_mail_domain(raw_domain: str) -> str:
+    if is_valid_ipv4(raw_domain):
+        return f"[{raw_domain}]"
+    DomainValidator().validate_domain_re(raw_domain)
+    return raw_domain
 
 conftestdir = Path(__file__).parent
-
-
-def _is_ip(domain):
-    try:
-        ipaddress.ip_address(domain)
-        return True
-    except ValueError:
-        return False
 
 
 def pytest_configure(config):
@@ -60,11 +58,6 @@ def chatmail_config(pytestconfig):
 @pytest.fixture(scope="session")
 def maildomain(chatmail_config):
     return chatmail_config.mail_domain_bare
-
-
-@pytest.fixture(scope="session")
-def maildomain_deliverable(maildomain):
-    return format_mail_domain(maildomain)
 
 
 @pytest.fixture(scope="session")
@@ -349,7 +342,7 @@ class ChatmailACFactory:
             account = self.dc.add_account()
             domain_deliverable = format_mail_domain(domain)
             addr, password = self.gencreds(domain_deliverable)
-            if _is_ip(domain):
+            if is_valid_ipv4(domain):
                 # Use DCLOGIN scheme with explicit server hosts,
                 # matching how madmail presents its addresses to users.
                 qr = (
@@ -423,10 +416,10 @@ class Remote:
     def iter_output(self, logcmd="", ready=None):
         getjournal = "journalctl -f" if not logcmd else logcmd
         print(self.sshdomain)
-        match self.sshdomain:
-            case "@local": command = []
-            case "localhost": command = []
-            case _: command = ["ssh", f"root@{self.sshdomain}"]
+        if self.sshdomain in ("@local", "localhost"):
+            command = []
+        else:
+            command = ["ssh", f"root@{self.sshdomain}"]
         [command.append(arg) for arg in getjournal.split()]
         popen = subprocess.Popen(
             command,
